Move EditableField to separate file
